menu type

+++ Africanacity_Frontend/Administrator/src/app/administration/menu-types/menu-types.component.ts
@@ -70,10 +70,12 @@ export class MenuTypesComponent implements OnInit{
   }
 
   //Delete menu type
-  deleteMenuType(menu_TypeId: Number){
-    this.dataService.deleteMenuType(menu_TypeId).subscribe(result => {
-      this.deleteItem();
-    });
+  deleteMenuType(menu_TypeId: number | undefined) {
+    if (menu_TypeId !== undefined) {
+      this.dataService.deleteMenuType(menu_TypeId).subscribe(result => {
+        this.deleteItem();
+      });
+    }
   }
 
   applyFilter(event: Event) {
